+++ c18-32-ft-java-angular/src/main/java/tech/nocountry/c1832ftjavaangular/service/EmailService.java
@@ -1,85 +1,41 @@
 package tech.nocountry.c1832ftjavaangular.service;
 
-import com.mailersend.sdk.MailerSend;
 import com.mailersend.sdk.Recipient;
-import com.mailersend.sdk.emails.Email;
-import com.mailersend.sdk.exceptions.MailerSendException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.io.Resource;
-import org.springframework.core.io.ResourceLoader;
-import org.springframework.stereotype.Service;
 
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.util.Collection;
 
-@Service()
-public class EmailService {
+public interface EmailService {
 
-    @Value("${mailersend.api.key}")
-    private String apiKey;
-    private final ResourceLoader resourceLoader;
+    /**
+     * Sends an HTML email to the specified recipients with the given subject and message.
+     *
+     * @param recipients the recipients of the email
+     * @param subject    the subject of the email
+     * @param message    the HTML content of the email
+     * @return true if the email was sent successfully, false otherwise
+     */
+    boolean sendHtmlEmail(Collection<Recipient> recipients, String subject, String message);
 
-    Logger logger = LoggerFactory.getLogger(EmailService.class);
+    /**
+     * Sends a plain text email to the specified recipients with the given subject and message.
+     * If useDefaultTemplate is true, the message is formatted using a default template.
+     *
+     * @param recipients         the recipients of the email
+     * @param subject            the subject of the email
+     * @param message            the plain text content of the email
+     * @param useDefaultTemplate whether to use a default template for the message
+     * @return true if the email was sent successfully, false otherwise
+     */
+    boolean sendPlainEmail(Collection<Recipient> recipients, String subject, String message, boolean useDefaultTemplate);
 
-    public EmailService(ResourceLoader resourceLoader) {this.resourceLoader = resourceLoader;}
+    /**
+     * Sends a plain text email to the specified recipients with the given subject and message.
+     *
+     * @param recipients the recipients of the email
+     * @param subject    the subject of the email
+     * @param message    the plain text content of the email
+     * @return true if the email was sent successfully, false otherwise
+     */
+    boolean sendPlainEmail(Collection<Recipient> recipients, String subject, String message);
 
-    private boolean sendEmail(Collection<Recipient> recipients, String subject, String content, boolean isHtml) {
-//        if (!activeProfile.equals("prod")) {
-//            logger.warn("Emails can only be sent in production environment");
-//            return true;
-//        }
-
-        Email email = new Email();
-
-        email.setFrom("Viajes Grupales", "do-not-reply@ericramirezs.com");
-        email.AddRecipients(recipients.toArray(new Recipient[0]));
-
-        email.setSubject(subject);
-        if (isHtml) {
-            email.setHtml(content);
-        } else {
-            email.setPlain(content);
-        }
-
-        MailerSend ms = new MailerSend();
-
-        System.out.println(apiKey);
-        ms.setToken(apiKey);
-        try {
-            ms.emails().send(email);
-            logger.info("Email sent successfully with subject {}", subject);
-            return true;
-        } catch (MailerSendException e) {
-            logger.error(e.getMessage());
-            return false;
-        }
-    }
-
-    public boolean sendHtmlEmail(Collection<Recipient> recipients, String subject, String message) {
-        return sendEmail(recipients, subject, message, true);
-    }
-
-    public boolean sendPlainEmail(Collection<Recipient> recipients, String subject, String message,
-                                  boolean useDefaultTemplate) {
-        if (useDefaultTemplate) {
-            try {
-                Resource resource = resourceLoader.getResource("classpath:template/EmailTemplate.html");
-                String template = resource.getContentAsString(StandardCharsets.UTF_8);
-                message = String.format(template, subject, message);
-            } catch (IOException e) {
-                logger.error(e.getMessage(), e);
-            }
-            return sendHtmlEmail(recipients, subject, message);
-        } else {
-            return sendEmail(recipients, subject, message, false);
-        }
-    }
-
-    public boolean sendPlainEmail(Collection<Recipient> recipients, String subject, String message) {
-        return sendEmail(recipients, subject, message, false);
-    }
 }
-
